APP_autoML/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==3.0.1

# Database
psycopg==3.2.5  # Pour PostgreSQL - compatible Python 3.13
SQLAlchemy==2.0.23

# Google OAuth2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# ML Libraries
xgboost==2.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.0.2
joblib==1.4.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0

# Cloud Storage (Production)
boto3==1.34.42  # AWS S3 support for model storage

# Development
pytest==7.4.3
pytest-flask==1.3.0
